Classes and oop (#3)
* Added the Cypher class for A4.1
This includes a .h and .cpp files with uncompleted implementation

* Added new build task
This includes all .cpp files in the call directory (works on Linux only)

* Finished A4.1
Implemented the rest of methods and added documentation

* Moved A4.1 files into separated folder

* Added solution to A4.2

* Minor formatting fixes and documentation addition to A4.1

* Added solution to A4.3 w/o CLI

* Extracted SIMcard class to separate files
Also removed unnecessary include guard in Caesar.h

* Minor formatting fixes

* Added SimpleCLI class
This is not included into main .cpp file yet

* Improved CLI logging methods
These now have an option to add a blank line after the message.

* Replaced SIMcard search for-each loop with simpler index

* Implemented a simple SIMcard program
This uses the previously implemented SimpleCLI tool.
Note that for now the search function doesn't work properly because the contact list remains empty after successful addition of new contacts T_T

* Fixed SIMcard contact list management
The bug was caused because of misunderstanding of stack memory. Now the program is able to save and search contacts correctly.
There's one funny bug: if a contact has the '0' number, it 'cannot be found' on the list because of overlapping with the default return value.

* Added notes about classes
Further TODO chapters include differences between creating objects in stack and heap

* Changed the chapter link to C++ in containers notes

* Added stack vs heap section

* Removed compiler command comment from SimpleCLI.h